waqi-collector/requirements.txt
requests
google-cloud-storage
google-cloud-secret-manager
Flask